Move ClassFinder to the fx module, which is the only place it's used.
